fix. Fikk ikke patchet til DVH
* More than one TaskExecutor bean found within the context, and none is named 'taskExecutor'. Mark one of them as primary or name it 'taskExecutor' (possibly as an alias) in order to use it for async processing: [arenaThreadPoolExecutor, taskScheduler]
