+++ src/main/java/no/nav/tag/tiltaksgjennomforing/datavarehus/DvhAvtalePatchService.java
@@ -6,7 +6,6 @@ import no.nav.tag.tiltaksgjennomforing.avtale.Avtale;
 import no.nav.tag.tiltaksgjennomforing.avtale.AvtaleRepository;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -17,7 +16,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 public class DvhAvtalePatchService {
 
     private final AvtaleRepository avtaleRepository;
-    private final DvhMeldingEntitetRepository dvhRepository;
+    private final DvhAvtalePatcher dvhAvtalePatcher;
 
     @Async
     public void lagDvhPatchMeldingForAlleAvtaler() {
@@ -25,28 +24,22 @@ public class DvhAvtalePatchService {
         List<Avtale> alleAvtaler = avtaleRepository.findAllByGjeldendeInnhold_AvtaleInngåttNotNull();
 
         alleAvtaler.forEach(avtale -> {
-            if(skalPatches(avtale)) {
-                lagDvhPatchMelding(avtale);
+            if (!skalPatches(avtale)) {
+                log.info("Avtale {} skal ikke patches i DVH", avtale.getId());
+            } else {
+                dvhAvtalePatcher.lagDvhPatchMelding(avtale);
                 antallPatchet.getAndIncrement();
-                if(antallPatchet.get() % 100 == 0) {
+                if (antallPatchet.get() % 100 == 0) {
                     log.info("Migrert {} antall avtaler", antallPatchet.get());
                 }
             }
-            log.info("Avtale {} skal ikke patches i DVH", avtale.getId());
         });
         log.info("Migrert {} antall avtaler", antallPatchet.get());
     }
 
-    @Transactional
-    public void lagDvhPatchMelding(Avtale avtale) {
-        var melding = AvroTiltakHendelseFabrikk.konstruer(avtale, DvhHendelseType.PATCHING, "system");
-        DvhMeldingEntitet entitet = new DvhMeldingEntitet(avtale, melding);
-        dvhRepository.save(entitet);
-    }
-
     private boolean skalPatches(Avtale avtale) {
-        if(avtale.erAvtaleInngått()) {
-            if(!avtale.erGodkjentAvVeileder()) {
+        if (avtale.erAvtaleInngått()) {
+            if (!avtale.erGodkjentAvVeileder()) {
                 log.warn("Avtale {} er inngått men ikke godkjent av veileder", avtale.getId());
                 return false;
             }
